+++ src/include/index/ivf_pq_group.h
@@ -84,7 +84,7 @@ class ivf_pq_group : public base_index_group<index_type> {
       tiledb_query_type_t rw = TILEDB_READ,
       TemporalPolicy temporal_policy = TemporalPolicy{TimeTravel, 0},
       const std::string& version = std::string{""},
-      uint64_t dimensions = 0,
+      uint32_t dimensions = 0,
       size_t num_clusters = 0,
       size_t num_subspaces = 0)
       : Base(ctx, uri, rw, temporal_policy, version, dimensions) {
